+++ lib/Perl/Tidy/VerticalAligner/Alignment.pm
@@ -9,7 +9,7 @@ package Perl::Tidy::VerticalAligner::Alignment;
 use strict;
 use warnings;
 
-our $VERSION = '20250912.02';
+our $VERSION = '20250912.03';
 
 sub new {